requirements.txt
azure-common==1.1.28
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.7.1
idna==3.10
jira==3.8.0
numpy==2.0.2
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
